Added drawer navigation, templates for new pages
