+++ alderman_spending/lib/l10n/app_en.arb
@@ -22,5 +22,21 @@
   "projectTitle":"Alderman Spending",
   "@projectTitle": {
     "description": "Won't work for the MaterialApp title property, would be used for the title of the project"
+  },
+  "home": "Home",
+  "@home": {
+    "description": "To be used as the label for the home button"
+  },
+  "faq": "FAQ",
+  "@faq": {
+    "description": "To be used as the label for the faq button"
+  },
+  "about": "About",
+  "@about": {
+    "description": "To be used as the label for the about button"
+  },
+  "spendingCharts": "Spending Charts",
+  "@spendingCharts": {
+    "description": "To be used as the label for the spending charts button"
   }
 }
