+++ alderman_spending/lib/l10n/app_es.arb
@@ -3,6 +3,10 @@
     "chartTitle": "Gasto Por Categoría",
     "detailPlaceholder": "Seleccione un barrio para ver los detalles de gasto",
     "errorLoadingData": "Error al cargar datos",
-    "noDataForWardYear":"No hay datos por el barrio {wardValue} en {yearValue}",
-    "projectTitle": "Gastos De Los Concejales"
+    "noDataForWardYear": "No hay datos por el barrio {wardValue} en {yearValue}",
+    "projectTitle": "Gastos De Los Concejales",
+    "home": "Inicio",
+    "faq": "Preguntas Frecuentes",
+    "about": "Acerca De",
+    "spendingCharts": "Gráficos De Gastos"
 }
