+++ alderman_spending/lib/src/ui/about_page/about_screen.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class AboutScreen extends StatelessWidget {
+  const AboutScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Placeholder();
+  }
+}
