+++ alderman_spending/lib/src/ui/faq_page/faq_screen.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class FAQScreen extends StatelessWidget {
+  const FAQScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Placeholder();
+  }
+}
