+++ alderman_spending/lib/src/ui/home_screen/home_screen.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class HomeScreen extends StatelessWidget {
+  const HomeScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Placeholder();
+  }
+}
