+++ alderman_spending/pubspec.yaml
@@ -68,8 +68,8 @@ flutter:
   assets:
     - assets/2019-2022_ward_category_totals.csv
     - assets/2019-2022_ward_items.csv
-  #   - images/a_dot_burr.jpeg
-  #   - images/a_dot_ham.jpeg
+    - assets/images/usa.png
+    - assets/images/mexico.png
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware
